% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCRG_function.R
\name{CRGs_extraction}
\alias{CRGs_extraction}
\title{Extraction of cooperation response genes and differentially expressed genes based on transcriptomic profile}
\usage{
CRGs_extraction()
}
\description{
This function allows you to extract cooperation response genes and differentially expressed genes.
}
\examples{
CRGs_extraction()
}

