% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCRG_function.R
\name{CRGs_extraction_inf}
\alias{CRGs_extraction_inf}
\title{Extraction of cooperation response genes and Network inference}
\usage{
CRGs_extraction_inf()
}
\description{
This function allows you to extract cooperation response genes and inference related the gene regulatory network.
}
\examples{
\donttest{
CRGs_extraction_inf()
}
}
