% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differentially_expressed_genes.R
\docType{data}
\name{Differentially_expressed_genes}
\alias{Differentially_expressed_genes}
\title{Differentially_expressed_genes.}
\format{data frame}
\usage{
Differentially_expressed_genes
}
\description{
List of differentially expressed genes in murine colon cancer model.
}
\examples{
data(Differentially_expressed_genes)
}
\keyword{datasets}

