% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Differentially_expressed_genes.R
\docType{data}
\name{{Differentially_expressed_genes}}
\alias{Differentially_expressed_genes}
\alias{{Differentially_expressed_genes}}
\title{{List of differentially expressed genes in a murine colon cancer model under single and combined mutations of mp53 and Ras}}
\format{An object of class \code{data.frame} with 540 rows and 4 columns.}
\usage{
{
data(Differentially_expressed_genes)
}
}
\description{
{
List of differentially expressed genes measured in young adult mouse colonic epithelium (YAMC) in four conditions: bleo/Neo vector control, mu-tant p53-expressing (mp53), activated Ras-expressing, and both mutant genes (mp53/Ras) cells using Affymetrix GeneChip® Mouse Genome 430 2.0 Arrays (GSE9199) (PMC2613942). 
}
}
\keyword{datasets}

