% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLCRG.R
\docType{data}
\name{{PLCRG}}
\alias{PLCRG}
\alias{{PLCRG}}
\title{{List of synergistic response genes in a murine colon cancer model under single and combined mutations of mp53 and Ras}}
\format{An object of class \code{data.frame} with 85 rows and 2 columns.}
\usage{
{
data(PLCRG)
}
}
\description{
{
List of differentially expressed and synergistic response genes measured in young adult mouse colonic epithelium (YAMC) in four conditions: bleo/Neo vector control, mu-tant p53-expressing (mp53), activated Ras-expressing, and both mutant genes (mp53/Ras) cells using Affymetrix GeneChip® Mouse Genome 430 2.0 Arrays (GSE9199) (PMC2613942). 
}
}
\keyword{datasets}

