% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCRG_function.R
\name{PL}
\alias{PL}
\title{Use the predefinded list of CRGs for Network inference}
\usage{
PL()
}
\description{
This function allows you to use the list of known cooperation response genes as input.
}
\examples{
\donttest{
PL()
}
}
