% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLCRG.R
\docType{data}
\name{PLCRG}
\alias{PLCRG}
\title{PLCRG.}
\format{data frame}
\usage{
PLCRG
}
\description{
List of cooperation response genes
least 5 uses.
}
\examples{
data(PLCRG)
}
\keyword{datasets}

