% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRGnet_function.R
\name{{SRG}}
\alias{SRG}
\alias{{SRG}}
\title{{Identification of differentially expressed and synergistic response genes (SRGs) in transcriptomics profile}}
\usage{
{
SRG(threshold of pvalue = 0.01)
}
}
\description{
{
We use the "SRG" function for identification of synergy or interaction effects between genes in transcriptomics profile in response to combination of mutations, drugs or environmental exposure. SRG returns lists of synergistic response genes and differentially expressed genes, which can be found in home directory of package as text file under title of "List_SRGs" and "List_DEGs", respectively.
}
}
\examples{
{
#using transcriptomic profiles from a murine colon cancer model
data(Transcriptomics)
SRG(0.01)
}
}
\author{
{Matthew McCall, Isar Nassiri}
}

