% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcriptomics.R
\docType{data}
\name{Transcriptomics}
\alias{Transcriptomics}
\title{Transcriptomics.}
\format{data frame}
\usage{
Transcriptomics
}
\description{
Transcriptomic profiles from a murine colon cancer model.
}
\keyword{datasets}