% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcriptomics.R
\docType{data}
\name{Transcriptomics}
\alias{Transcriptomics}
\title{{Murine colon cancer transcriptomics profile under single and combined mutations of mp53 and Ras}}
\format{An object of class \code{data.frame} with 483 rows and 41 columns.}
\usage{
Transcriptomics
}
\description{
{
Young adult mouse colonic epithelium (YAMC) transcriptomics profile in four conditions: YAMC control, mu-tant p53-expressing (mp53), activated Ras-expressing, and both mutant genes (mp53/Ras) cells using Affymetrix GeneChip® Mouse Genome 430 2.0 Arrays (GSE9199) (PMC2613942). 
}
}
\keyword{datasets}

